requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.6
django-appconf==1.0.4
django-cors-headers==3.3.0
django-crispy-forms==1.9.1
django-select2==7.4.2
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
docutils==0.16
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
inflection==0.5.0
Kivy==1.11.1
kivy-deps.angle==0.1.10
kivy-deps.glew==0.1.12
kivy-deps.gstreamer==0.1.18
kivy-deps.sdl2==0.1.23
Kivy-examples==1.11.1
Kivy-Garden==0.1.4
msgpack==0.6.2
Pillow==7.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pypiwin32==223
pytz==2018.5
pywin32==227
redis==2.10.6
requests==2.23.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
websockets==8.1
zope.interface==5.1.0
